refactor(web-app): Split nav component to auth module
Added hooks to common directory, start of modules dir. And some js-ts moves
