+++ src/customComponents/ProjectProfile/components/SubmitReviewTx.tsx
@@ -2,9 +2,9 @@ import { EventRecord } from '@polkadot/types/interfaces';
 import { useEffect, useState } from 'react';
 import { Button, Container, Header } from 'semantic-ui-react';
 import AccountSelector from '../../../AccountSelector';
+import { useLoadAccounts } from '../../../modules/Auth/hooks/useLoadAccounts';
 import { useSubstrate } from '../../../substrate-lib';
 import { TxButton } from '../../../substrate-lib/components';
-import { useLoadAccounts } from '../../menuBar';
 import { useApp } from '../../state';
 import { useReviewSend } from '../hooks/useReviewSend';
 import { EventView } from './EventView';
